Add files via upload
So these are the files for the cash register. We still need to do a couple of things to finish it up. 1) Make sure you understand these code files, how they work and how they go together. 2) Complete the input file and output file function and implement it. 3) Improve upon the existing code to  make it more efficient since we are graded on how efficient and clean our code is. 
If anyone has something else they'd like to add to the list of things to complete please do it in a comment. And if you take out the code, make sure its on a new branch and DO NOT commit it to the master branch until at least one other person has test run it. And if you are test running it, your job is to break the code, which also means you don't necessarily have to follow the on screen prompts given. Like if it asks for an int try inputting a char and see what happens. That's about it. If you don't understand the code, please ask because if someone knows then we'll all know and if no one knows, then we know we h